Enabled AJAX interactions with database
As it stands now, we have a separate post.js file that is used to create each Post (so the post object function we have in this file isn't really needed at the moment, but I left it in case there's code in it that someone wants to use for future reference).  For the timestamp, it's currently formatted to hold yyyy-mm-dd dates, so the function we have now that includes hours and minutes and seconds is too specific; I don't know if SQL can hold dates/times with that much precision, but if it can we'll have to alter the database schema to account for this (which should be easy).  Also, the JSON "message" value that we're getting back from the server doesn't seem to have newlines, or the way we're formatting it in html or js or css is incorrect, so we'll have to fix that, too.